rename: modal별 폴더 생성
